move custom captcha command
